requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==4.0.3
django-debug-toolbar==3.2.4
djlint==0.7.6
idna==3.3
importlib-metadata==4.11.3
pathspec==0.9.0
Pillow==9.0.1
pycodestyle==2.8.0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sqlparse==0.4.2
stripe==2.71.0
toml==0.10.2
tomlkit==0.9.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
zipp==3.8.0
